+++ script/init_websales.php
@@ -117,12 +117,14 @@ if (isset($_FILES['importfile']) && $_FILES['importfile']['error'] == UPLOAD_ERR
 						$socid = $obj->rowid; // Récupération de l'ID de la société
 					} else $socid = $fk_company;
 
-					$result = $actionsDolistore->addWebmoduleSales($TItemDatas, $socid);
+					if(isModEnabled("webhost")) {
+						$result = $actionsDolistore->addWebmoduleSales($TItemDatas, $socid);
 
-					if ($result <= 0) {
-						// Si l'insertion échoue, incrémenter l'erreur et afficher un message d'erreur
-						$error++;
-						echo displayStyledMessage($langs->transnoentitiesnoconv("FailedToInsertDataFor") . ' ' . htmlspecialchars($TItemDatas['item_reference'])  . $actionsDolistore->logCat.', '.$row[5], "error");
+						if ($result <= 0) {
+							// Si l'insertion échoue, incrémenter l'erreur et afficher un message d'erreur
+							$error++;
+							echo displayStyledMessage($langs->transnoentitiesnoconv("FailedToInsertDataFor") . ' ' . htmlspecialchars($TItemDatas['item_reference'])  . $actionsDolistore->logCat.', '.$row[5], "error");
+						}
 					}
 				}
 			}