+++ ChangeLog.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+## Release 1.5 (CAP-REL)
 
+- Compat Dolibarr 18
+- Compat with dolistore 2025
 
 ## Release 1.4
 - FIX : Compat V20 & V21 - *24/02/2025* - 1.4.3
@@ -21,6 +24,6 @@
 - FIX : tâche cron dolistorextract => fix gestion des erreurs - 1.2.10 - *28/03/2023*
 - FIX : Import contact problem for V15  - 1.2.9 - *24/06/2022*
 - FIX : everyday same thirdparty creation because of contact fetch return 2 if several matches, now if several matches we use the oldest thirdparty - 1.2.8 - *06/10/2021*
-- FIX : fails to create thirdparty / contact when more than one third party name matches 
+- FIX : fails to create thirdparty / contact when more than one third party name matches
   name from e-mail - 1.2.7 - *10/05/2021*
 - no ChangeLog up to this point